Revert "IComposerClient wrapper for VTS tests"
This reverts commit 634340d6a8e34c3688227b88ba8ccde341076b2b.

Reason for revert: No one is using it, I will make a cl with usage of the wrapper along side this and reupload it.

+++ graphics/composer/aidl/include/android/hardware/graphics/composer3/ComposerClientWriter.h
@@ -189,7 +189,7 @@ class ComposerClientWriter {
 
     void setLayerZOrder(int64_t display, int64_t layer, uint32_t z) {
         ZOrder zorder;
-        zorder.z = static_cast<int32_t>(z);
+        zorder.z = z;
         getLayerCommand(display, layer).z.emplace(std::move(zorder));
     }
 
@@ -228,9 +228,9 @@ class ComposerClientWriter {
     std::optional<LayerCommand> mLayerCommand;
     std::vector<DisplayCommand> mCommands;
 
-    Buffer getBuffer(uint32_t slot, const native_handle_t* bufferHandle, int fence) {
+    Buffer getBuffer(int slot, const native_handle_t* bufferHandle, int fence) {
         Buffer bufferCommand;
-        bufferCommand.slot = static_cast<int32_t>(slot);
+        bufferCommand.slot = slot;
         if (bufferHandle) bufferCommand.handle.emplace(::android::dupToAidl(bufferHandle));
         if (fence > 0) bufferCommand.fence = ::ndk::ScopedFileDescriptor(fence);
         return bufferCommand;
